+++ src/Command/ResetCustomSettingsCommand.php
@@ -57,7 +57,9 @@ class ResetCustomSettingsCommand extends Command
             ['name' => 'WELCOME_TEXT', 'value' => 'Welcome to OpenRoaming Provisioning Service'],
             [
                 'name' => 'WELCOME_DESCRIPTION',
-                'value' => 'This provisioning portal is for the WBA OpenRoaming Live Program'
+                // phpcs:disable Generic.Files.LineLength.TooLong
+                'value' => 'This portal allows you to download and install an OpenRoaming profile tailored to your device, allowing you to connect automatically to OpenRoaming Wi-Fi networks across the world.',
+                // phpcs:enable
             ],
             ['name' => 'ADDITIONAL_LABEL', 'value' => 'This label is used to add extra content if necessary'],
             ['name' => 'CONTACT_EMAIL', 'value' => 'openroaming-help@example.com'],